chore: move filter and lp tests to subdir
