+++ packages/tests/tests/filter/single_node/multiple_pubsub.node.spec.ts
@@ -20,9 +20,8 @@ import {
   MessageCollector,
   ServiceNode,
   tearDownNodes
-} from "../../src/index.js";
-
-import { runNodes } from "./utils.js";
+} from "../../../src/index.js";
+import { runNodes } from "../utils.js";
 
 describe("Waku Filter V2: Multiple PubsubTopics", function () {
   // Set the timeout for all tests in this suite. Can be overwritten at test level
