+++ packages/tests/tests/filter/single_node/ping.node.spec.ts
@@ -7,15 +7,14 @@ import {
   MessageCollector,
   ServiceNode,
   tearDownNodes
-} from "../../src/index.js";
-
+} from "../../../src/index.js";
 import {
   runNodes,
   TestContentTopic,
   TestDecoder,
   TestEncoder,
   validatePingError
-} from "./utils.js";
+} from "../utils.js";
 
 describe("Waku Filter V2: Ping", function () {
   // Set the timeout for all tests in this suite. Can be overwritten at test level
