+++ packages/tests/tests/filter/single_node/push.node.spec.ts
@@ -12,15 +12,14 @@ import {
   tearDownNodes,
   TEST_STRING,
   TEST_TIMESTAMPS
-} from "../../src/index.js";
-
+} from "../../../src/index.js";
 import {
   messageText,
   runNodes,
   TestContentTopic,
   TestDecoder,
   TestEncoder
-} from "./utils.js";
+} from "../utils.js";
 
 describe("Waku Filter V2: FilterPush", function () {
   // Set the timeout for all tests in this suite. Can be overwritten at test level
