+++ packages/tests/tests/filter/single_node/subscribe.node.spec.ts
@@ -20,8 +20,7 @@ import {
   ServiceNode,
   tearDownNodes,
   TEST_STRING
-} from "../../src/index.js";
-
+} from "../../../src/index.js";
 import {
   messagePayload,
   messageText,
@@ -29,7 +28,7 @@ import {
   TestContentTopic,
   TestDecoder,
   TestEncoder
-} from "./utils.js";
+} from "../utils.js";
 
 describe("Waku Filter V2: Subscribe", function () {
   // Set the timeout for all tests in this suite. Can be overwritten at test level
