+++ packages/tests/tests/filter/single_node/unsubscribe.node.spec.ts
@@ -9,8 +9,7 @@ import {
   MessageCollector,
   ServiceNode,
   tearDownNodes
-} from "../../src/index.js";
-
+} from "../../../src/index.js";
 import {
   messagePayload,
   messageText,
@@ -18,7 +17,7 @@ import {
   TestContentTopic,
   TestDecoder,
   TestEncoder
-} from "./utils.js";
+} from "../utils.js";
 
 describe("Waku Filter V2: Unsubscribe", function () {
   // Set the timeout for all tests in this suite. Can be overwritten at test level
